+++ dataset.csv
@@ -1,0 +1,11 @@
+chroma_stft,spectral_centroid,spectral_bandwidth,rolloff,zero_crossing_rate,mfcc1,mfcc2,mfcc3,mfcc4,mfcc5,mfcc6,mfcc7,mfcc8,mfcc9,mfcc10,mfcc11,mfcc12,mfcc13,mfcc14,mfcc15,mfcc16,mfcc17,mfcc18,mfcc19,mfcc20,label
+0.3689647614955902,1913.3871500955263,2085.802609039251,3969.2733457424497,0.09656873576821669,-330.7585754394531,96.4792709350586,25.496858596801758,21.337900161743164,-15.437956809997559,-4.430291652679443,-13.490593910217285,-1.190808892250061,-13.977445602416992,0.8758944272994995,-18.6108341217041,3.1869056224823,-7.726667404174805,-10.97693920135498,7.927337169647217,-8.421019554138184,-4.337893009185791,-5.667142868041992,-4.713412284851074,-5.2927327156066895,unsober
+0.3617429733276367,2338.1211290724973,2213.644037179328,4709.503845506878,0.13949891596889952,-387.0799865722656,90.51492309570312,12.857476234436035,23.81073760986328,-19.69330406188965,-0.24884437024593353,-12.63005542755127,-5.692428112030029,-21.33329200744629,14.504405975341797,-22.65970802307129,3.063835620880127,-8.131794929504395,-7.221322059631348,7.093003273010254,-12.2362060546875,0.7515645027160645,-4.031706809997559,-3.8383572101593018,-5.358673095703125,unsober
+0.31622931361198425,2094.7682326808545,2093.330553674339,4232.928028876471,0.11881437429932735,-362.0502014160156,103.84007263183594,6.562873363494873,26.327560424804688,-18.90403175354004,-10.083721160888672,-25.038944244384766,-1.8876969814300537,-13.194986343383789,13.794629096984863,-25.396760940551758,5.014101028442383,-6.077053070068359,-8.28077220916748,3.4475343227386475,-13.804792404174805,-0.949917733669281,-12.075712203979492,-16.093364715576172,-6.291658878326416,unsober
+0.3526739478111267,2084.4478172880854,2210.6197756045935,4245.9995314537155,0.10543087159794655,-397.83636474609375,101.1368179321289,20.00250244140625,20.427907943725586,-11.990020751953125,10.018077850341797,-15.497770309448242,1.1744604110717773,-21.213638305664062,15.170998573303223,-17.180498123168945,-1.8502647876739502,-9.146952629089355,-4.411040306091309,3.8137190341949463,-10.777970314025879,-0.14146608114242554,-4.550715446472168,-4.903192520141602,-1.8270777463912964,unsober
+0.31680217385292053,1503.2146637321323,1738.4308325671295,2986.2431498961746,0.0727689865319373,-300.0462951660156,116.60967254638672,12.152746200561523,15.995051383972168,-4.158257007598877,1.7333719730377197,-20.332883834838867,-17.206754684448242,-20.57204246520996,-2.330965995788574,-22.575075149536133,-6.711031436920166,-10.176379203796387,-9.404340744018555,-0.9287794828414917,-7.8602166175842285,2.3928771018981934,-6.738678455352783,-1.8719626665115356,-0.34826675057411194,unsober
+0.35035228729248047,2122.220231657273,2102.2154410721632,4145.504283792964,0.125442379369859,-310.76800537109375,102.9553451538086,17.59544563293457,31.119815826416016,-8.59702205657959,5.317625999450684,-29.70669937133789,-2.4296793937683105,-21.381881713867188,12.99862289428711,-18.888296127319336,-1.2841496467590332,-7.963425159454346,-12.204014778137207,0.7578730583190918,-7.819197177886963,-3.785250663757324,-10.398508071899414,-4.058253288269043,-7.450796127319336,unsober
+0.3500624895095825,1942.0586157449577,1996.8297208450824,3885.7042008032663,0.1031516031970538,-249.63796997070312,117.53050231933594,-6.71111536026001,13.122370719909668,-6.290326118469238,2.7529475688934326,-23.34392547607422,8.215460777282715,-18.325706481933594,-0.6842367053031921,-22.010374069213867,0.24514906108379364,-6.541111469268799,-14.324782371520996,-0.5547317266464233,-6.604910373687744,-5.451412200927734,-1.5141456127166748,3.4690167903900146,-4.377875328063965,unsober
+0.325242817401886,2056.348469554593,2086.3763447767983,4282.132439864309,0.10478071732954546,-361.3418884277344,97.8098373413086,9.164214134216309,35.08882141113281,-14.58597469329834,-11.027320861816406,-15.290693283081055,-6.952499866485596,-28.65005874633789,7.516440391540527,-20.65555191040039,-0.3250870406627655,-7.4462809562683105,-10.313498497009277,5.400264263153076,-12.745476722717285,-9.249576568603516,-2.270836591720581,-5.05482292175293,-7.091762065887451,unsober
+0.33576732873916626,1787.6968721577828,2029.9048045421532,3640.5573678049336,0.08660891460141618,-267.169189453125,110.74524688720703,15.263574600219727,17.331026077270508,-16.067466735839844,-3.860063314437866,-22.725481033325195,-5.824465751647949,-31.25798797607422,3.051316022872925,-21.66124725341797,-1.389609456062317,-1.4956121444702148,-15.241804122924805,1.0031055212020874,-3.455277442932129,-7.264087677001953,-1.9963679313659668,0.776394248008728,-8.836950302124023,unsober
+0.38492605090141296,1910.347210835174,2100.6450049871682,3932.9537990873896,0.10174206458947886,-429.45904541015625,111.15471649169922,31.186689376831055,34.77750778198242,-9.81762981414795,0.15024295449256897,-9.783951759338379,-4.498025894165039,-21.416492462158203,11.307316780090332,-18.523670196533203,0.936448872089386,-5.416940212249756,-9.839784622192383,3.8322322368621826,-11.654243469238281,1.7632453441619873,-6.949851989746094,-3.766326427459717,-4.710940361022949,unsober
