+++ sobriety_detection/main.py
@@ -16,11 +16,13 @@ Second class --> steps: 4.
 Third class --> steps: 6.
 Fourth class --> steps: 7, 8.
 '''
+
+from data_operator import Data_operator
+
 def main():
-
-
-    pass
-
+    
+    obj = Data_operator('data_test2')
+    obj.import_audio()
 
 if __name__ == '__main__':
     main()
